modificando el delete de servicio para validacion

+++ demo/src/main/java/com/example/demo/service/PeliculaServiceImpl.java
@@ -39,7 +39,11 @@ public class PeliculaServiceImpl implements PeliculaService {
         return null;
     }
     @Override
-    public void deletePelicula(Long id){
-        peliculaRep.deleteById(id);
+    public Boolean deletePelicula(Long id){
+        if(peliculaRep.existsById(id)){
+            peliculaRep.deleteById(id);
+            return true;
+        }
+        return false;
     }
 }
